Fix list array vs list discrepancies (#142)
### Checklist

* [x] I have read the [Contributor Guide](../../CONTRIBUTING.md)
* [x] I have read and agree to the [Code of
Conduct](../../CODE_OF_CONDUCT.md)
* [x] I have added a description of my changes and why I'd like them
included in the section below

### Description of Changes

This fixes a few issues with arrays that I discovered:

- `<[T; N]>::from_reflect` for `Value::List` would previously fail. It
now works if the lengths match.
- `Vec::as_array`, `Vec::as_array_mut`, `Vec::into_array` now correctly
return `Some(_)`.
- `Value::as_array`, `Value::as_array_mut`, `Value::into_array` now
correctly return `Some(_)` for `Value::List`.

Basically lists are a subtype of arrays so if something is a list then
its also an array, i.e. `as_array` should work on a `&dyn List`.