Update `kollect` to 0.4.0 (#144)
This bumps `kollect` to 0.4.0, which I've just released, the only change
being changing the default serialization format of maps to be as
sequences of `(k, v)` pairs. Thus this is a breaking serialization
change. We should therefore not bump this version into wim until the PR
that intentionally breaks a bunch of data which this is being used for.